README.md
This project investigated i) The breakdown of aflatoxin B1 and G1 by aflatoxin-producing (toxigenic) and non-aflatoxin-producing (atoxigenic) Flavi. (Aspergillus section Flavi) ii) Sporulation in Flavi with the influence of an antioxidative environment. 
Aflatoxin is known to play a role in oxidative stress alleviation in Flavi. Some species of Flavi are also known to degrade aflatoxin. Hence, its would be helpful to understand how an antioxidative environment would influence the production of aflatoxin in Flavi as well as its degradation. Secondly, differences in Darwinian fitness between toxigenic and atoxigenic Flavi has potential to determine the population dynamics between the two variants of Flavi.